Reorganize MIF modules.
