+++ src/modelarrayio/cli/h5_to_mif.py
@@ -51,7 +51,6 @@ def h5_to_mif(example_mif, in_file, analysis_name, output_dir):
     """
     # Use the example MIF as the template so layout and metadata stay native to MIF.
     template_img, _ = mif_to_image(example_mif)
-    template_shape = template_img.shape
     output_path = Path(output_dir)
     with h5py.File(in_file, 'r') as h5_data:
         results_matrix = h5_data[f'results/{analysis_name}/results_matrix']
@@ -62,41 +61,51 @@ def h5_to_mif(example_mif, in_file, analysis_name, output_dir):
         for result_col, result_name in enumerate(results_names):
             valid_result_name = cli_utils.sanitize_result_name(result_name)
             out_mif = output_path / f'{analysis_name}_{valid_result_name}.mif'
-            if out_mif.exists():
-                logger.warning('Output file already exists. Not overwriting. %s', out_mif)
-                continue
-
-            result_data = np.asarray(results_matrix[result_col, :], dtype=np.float32).reshape(
-                template_shape
+            write_mif(
+                arr=np.asarray(results_matrix[result_col, :], dtype=np.float32),
+                template_img=template_img,
+                out_file=out_mif,
             )
-            result_header = template_img.header.copy()
-            result_header.set_data_shape(result_data.shape)
-            result_header.set_data_dtype(result_data.dtype)
-            result_img = MifImage(result_data, template_img.affine, header=result_header)
-            result_img.to_filename(out_mif)
 
             if 'p.value' not in valid_result_name:
                 continue
 
             valid_result_name_1mpvalue = valid_result_name.replace('p.value', '1m.p.value')
             out_mif_1mpvalue = output_path / f'{analysis_name}_{valid_result_name_1mpvalue}.mif'
-            if out_mif_1mpvalue.exists():
-                logger.warning('Output file already exists. Not overwriting. %s', out_mif_1mpvalue)
-                continue
-
             output_mifvalues_1mpvalue = np.asarray(
                 1 - results_matrix[result_col, :],
                 dtype=np.float32,
-            ).reshape(template_shape)
-            output_header_1mpvalue = template_img.header.copy()
-            output_header_1mpvalue.set_data_shape(output_mifvalues_1mpvalue.shape)
-            output_header_1mpvalue.set_data_dtype(output_mifvalues_1mpvalue.dtype)
-            output_img_1mpvalue = MifImage(
-                output_mifvalues_1mpvalue,
-                template_img.affine,
-                header=output_header_1mpvalue,
             )
-            output_img_1mpvalue.to_filename(out_mif_1mpvalue)
+            write_mif(
+                arr=output_mifvalues_1mpvalue,
+                template_img=template_img,
+                out_file=out_mif_1mpvalue,
+            )
+
+
+def write_mif(arr, template_img, out_file):
+    """Write array to MIF file.
+
+    Parameters
+    ----------
+    arr : :obj:`numpy.ndarray`
+        Array to write to file.
+    template_img : :obj:`MifImage`
+        Template MIF image.
+    out_file : :obj:`pathlib.Path`
+        Output file to write.
+        If it already exists, this function will raise a warning and not overwrite.
+    """
+    if out_file.exists():
+        logger.warning('Output file already exists. Not overwriting. %s', out_file)
+        return
+
+    result_data = arr.reshape(template_img.shape)
+    result_header = template_img.header.copy()
+    result_header.set_data_shape(result_data.shape)
+    result_header.set_data_dtype(result_data.dtype)
+    result_img = MifImage(result_data, template_img.affine, header=result_header)
+    result_img.to_filename(out_file)
 
 
 def h5_to_mif_main(
